backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.1.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
